Add tournament pubsub & use ecto DSL
