+++ lib/lanpartyseating/logic/tournaments_logic.ex
@@ -1,16 +1,18 @@
 defmodule Lanpartyseating.TournamentsLogic do
   import Ecto.Query
-  alias Lanpartyseating.PubSub, as: PubSub
-  alias Lanpartyseating.StationLogic, as: StationLogic
-  alias Lanpartyseating.Tournament, as: Tournament
-  alias Lanpartyseating.Repo, as: Repo
+  alias Lanpartyseating.SettingsLogic
+  alias Lanpartyseating.PubSub
+  alias Lanpartyseating.StationLogic
+  alias Lanpartyseating.Tournament
+  alias Lanpartyseating.TournamentReservation
+  alias Lanpartyseating.Repo
 
   @spec get_all_tournaments :: any
   def get_all_tournaments do
-    Tournament
-    |> where([v], is_nil(v.deleted_at))
-    |> order_by([v], asc: v.end_date)
-    |> Repo.all()
+    from(t in Tournament,
+      where: is_nil(t.deleted_at),
+      order_by: [asc: t.end_date]
+    ) |> Repo.all()
   end
 
   @spec get_all_daily_tournaments :: any
@@ -21,31 +23,52 @@ defmodule Lanpartyseating.TournamentsLogic do
     time = elem(Time.new(4, 0, 0, 0), 1)
     tomorrow = elem(DateTime.new(date, time, "Etc/UTC"), 1)
 
-    Tournament
-    |> where([v], v.end_date > from_now(0, "second") and v.end_date < ^tomorrow)
-    |> where([v], is_nil(v.deleted_at))
-    |> Repo.all()
+    from(t in Tournament,
+      where: t.end_date > from_now(0, "second"),
+      where: t.end_date < ^tomorrow,
+      where: is_nil(t.deleted_at)
+    ) |> Repo.all()
   end
 
   def get_upcoming_tournaments do
-    Tournament
-    |> where([v], v.end_date > from_now(0, "second"))
-    |> where([v], is_nil(v.deleted_at))
-    |> Repo.all()
+    from(t in Tournament,
+      where: t.end_date > from_now(0, "second"),
+      where: is_nil(t.deleted_at)
+    ) |> Repo.all()
   end
 
   def create_tournament(name, start_time, duration) do
     end_time = DateTime.add(start_time, duration, :hour, Tzdata.TimeZoneDatabase)
 
     case Repo.insert(%Tournament{start_date: start_time, end_date: end_time, name: name}) do
-      {:ok, tournament} -> {:ok, tournament}
+      {:ok, tournament} ->
+        start_delay = DateTime.diff(tournament.start_date, DateTime.utc_now(), :millisecond)
+        expiry_delay = DateTime.diff(tournament.end_date, DateTime.utc_now(), :millisecond)
+
+        DynamicSupervisor.start_child(
+          Lanpartyseating.ExpirationTaskSupervisor,
+          {Lanpartyseating.Tasks.StartTournament, {start_delay, tournament.id}}
+        )
+
+        DynamicSupervisor.start_child(
+          Lanpartyseating.ExpirationTaskSupervisor,
+          {Lanpartyseating.Tasks.ExpireTournament, {expiry_delay, tournament.id}}
+        )
+
+        Phoenix.PubSub.broadcast(
+          PubSub,
+          "tournament_update",
+          {:tournaments, get_all_tournaments()}
+        )
+        {:ok, tournament}
     end
   end
 
   def delete_tournament(id) do
-    Tournament
-    |> where(id: ^id)
-    |> where([v], is_nil(v.deleted_at))
+    from(t in Tournament,
+      where: t.id == ^id,
+      where: is_nil(t.deleted_at)
+    )
     |> Repo.all()
     |> Enum.map(fn res ->
       tournament =
@@ -62,8 +85,54 @@ defmodule Lanpartyseating.TournamentsLogic do
             "station_update",
             {:stations, StationLogic.get_all_stations()}
           )
+          Phoenix.PubSub.broadcast(
+            PubSub,
+            "tournament_update",
+            {:tournaments, get_all_tournaments()}
+          )
           {:ok, struct}
       end
     end)
+  end
+
+  def create_tournament_reservations_by_range(start_station, end_station, tournament_id) do
+    # Input validation
+    settings = SettingsLogic.get_settings()
+
+    cond do
+      start_station < 1 or start_station > settings.columns * settings.rows ->
+        {:error, "Start station is out of bounds"}
+
+      end_station < 1 or end_station > settings.columns * settings.rows ->
+        {:error, "End station is out of bounds"}
+
+      start_station > end_station ->
+        {:error, "Start station is after end station"}
+
+      true ->
+        now = NaiveDateTime.truncate(NaiveDateTime.utc_now(), :second)
+        reservations = StationLogic.get_stations_by_range(
+          start_station,
+          end_station
+        )
+        |> Enum.map(fn station ->
+          %{
+            tournament_id: tournament_id,
+            station_id: station.id,
+            inserted_at: now,
+            updated_at: now
+          }
+        end)
+
+        inserted = Repo.insert_all(TournamentReservation, reservations)
+
+        Phoenix.PubSub.broadcast(
+          PubSub,
+          "station_update",
+          {:stations, StationLogic.get_all_stations()}
+        )
+
+        {:ok, inserted}
+    end
   end
 end
